DPG-1714: PR review fixes and clean up, typography integrated
