+++ frontend/micro-ui/web/micro-ui-internals/packages/components-core/src/atoms/Button.js
@@ -13,7 +13,9 @@ const Button = (props) => {
       style={props.style ? props.style : null}
     >
       {props?.icon && props.icon}
-      <h2 style={{ ...props?.textStyles, ...{ width: "100%" } }}>{props.label}</h2>
+      <h2 style={{ ...props?.textStyles }} className="digit-button-label">
+        {props.label}
+      </h2>
     </button>
   );
 };
