+++ frontend/micro-ui/web/micro-ui-internals/packages/components-core/src/atoms/ButtonSelector.js
@@ -20,7 +20,9 @@ const ButtonSelector = (props) => {
       disabled={props.isDisabled}
       style={props.style ? props.style : null}
     >
-      <h2 style={{ ...props?.textStyles, ...{ width: "100%" } }}>{props.label}</h2>
+      <h2 style={{ ...props?.textStyles }} className="digit-button-label">
+        {props.label}
+      </h2>
     </button>
   );
 };
