+++ frontend/micro-ui/web/micro-ui-internals/packages/css/src/digitv2/components/FormComposerV2.scss
@@ -17,10 +17,10 @@
 }
 
 .card-date-input {
+  @extend .light-input-border;
   @apply absolute top-0 left-0 w-3/4 bg-white pl-sm border border-input-border border-solid border-r-0;
   outline: transparent solid 1px;
   height: 2.5rem;
-  @extend .light-input-border;
 }
 .text-input-width {
   max-width: 540px;
@@ -31,59 +31,59 @@
   input {
     background-color: transparent;
     &:hover {
-      @apply border-2 border-solid border-primary-main;
       @extend .light-input-border;
+      @apply border-2 border-solid border-primary-main;
     }
   }
 }
 .employee-card-input-error {
-  @apply mb-lg pl-sm block w-full h-10 outline-none border-2 border-error border-solid bg-white leading-10 text-form-field text-text-primary;
   @extend .light-input-border;
   @extend .alert-error;
+  @apply mb-lg pl-sm block w-full h-10 outline-none border-2 border-error border-solid bg-white leading-10 text-form-field text-text-primary;
 }
 .card-input-error {
-  @apply mb-lg pl-sm block w-full h-10 outline-none border-2 border-error border-solid bg-white leading-10 text-form-field text-text-primary;
   @extend .light-input-border;
   @extend .alert-error;
+  @apply mb-lg pl-sm block w-full h-10 outline-none border-2 border-error border-solid bg-white leading-10 text-form-field text-text-primary;
 }
 
 .employee-card-input {
-  @apply mb-lg pl-sm outline-none block border w-full h-10 border-input-border border-solid bg-white leading-10 text-form-field text-text-primary;
   @extend .light-input-border;
+  @apply mb-lg pl-sm outline-none block border w-full h-10 border-input-border border-solid bg-white leading-10 text-form-field text-text-primary;
 }
 .employee-card-input:disabled {
-  @apply border-grey-dark text-grey-dark !important;
-  pointer-events: none !important;
   @extend .light-input-border;
   @extend .light-text-color-disabled;
+  @apply border-grey-dark text-grey-dark !important;
+  pointer-events: none !important;
 }
 .employee-card-input--front {
+  @extend .light-background;
   width: fit-content !important;
   display: flex;
   align-items: center;
   /* background: theme(colors.grey.mid); */
-  @extend .light-background;
   border-right: 0;
   padding-right: 5px;
 }
 .employee-card-input-error {
-  @apply mb-lg pl-sm block w-full h-10 outline-none border-2 border-error border-solid bg-white leading-10 text-form-field text-text-primary;
   @extend .light-input-border;
   @extend .alert-error;
+  @apply mb-lg pl-sm block w-full h-10 outline-none border-2 border-error border-solid bg-white leading-10 text-form-field text-text-primary;
 }
 
 .card-textarea,
 .employee-card-textarea {
-  @apply block outline-none mt-md mb-lg border w-full border-input-border border-solid bg-white h-24 text-form-field text-text-primary p-sm;
   @extend .light-input-border;
+  @apply block outline-none mt-md mb-lg border w-full border-input-border border-solid bg-white h-24 text-form-field text-text-primary p-sm;
 }
 .employee-card-textarea {
   @apply w-full;
 }
 
 .citizen-card-input {
-  @apply mb-lg pl-sm outline-none block border w-full h-10 border-input-border border-solid bg-white leading-10 text-form-field text-text-primary;
   @extend .light-input-border;
+  @apply mb-lg pl-sm outline-none block border w-full h-10 border-input-border border-solid bg-white leading-10 text-form-field text-text-primary;
 }
 .citizen-card-input--front {
   @extend .light-background;
@@ -101,19 +101,18 @@
 }
 
 .primary-label-btn {
+  @extend .light-primary;
   @apply flex;
+  cursor: pointer;
+  font-weight: 500;
+  width: fit-content;
   gap: 10px;
 
   svg {
     fill: theme(colors.primary.main);
   }
-
-  @extend .light-primary;
-  cursor: pointer;
-  font-weight: 500;
-  width: fit-content;
 }
 
-.digit-form-composer-header{
-   font-size:theme(digitv2.fontSize.heading-l-dt);
+.digit-form-composer-header {
+  font-size: theme(digitv2.fontSize.heading-l-dt);
 }
