+++ frontend/micro-ui/web/micro-ui-internals/packages/css/src/digitv2/components/actionbarV2.scss
@@ -14,20 +14,20 @@
   }
 
   .digit-menu-wrap {
-    @apply absolute bg-white text-left mb-xs z-30 cursor-pointer;
     @extend .light-background;
+    @apply absolute bg-white text-left mb-xs z-30 cursor-pointer;
     bottom: 45px;
     box-shadow: rgba(0, 0, 0, 0.14) 0 8px 10px 1px, rgba(0, 0, 0, 0.12) 0 3px 14px 2px, rgba(0, 0, 0, 0.2) 0 5px 5px -3px;
     width: calc(100% - 16px);
     right: 8px;
 
     div {
-      @apply h-12 pl-sm py-md text-body-l text-text-primary;
       @extend .light-primary;
+      @apply h-12 pl-sm py-md text-body-l text-text-primary;
 
       &:hover {
-        @apply bg-grey-mid w-full;
         @extend .light-background;
+        @apply bg-grey-mid w-full;
       }
     }
   }
@@ -72,25 +72,25 @@
     @apply w-full;
   }
   .digit-menu-wrap {
+    @extend .light-background;
     @apply absolute bg-white text-left mb-xs z-30 cursor-pointer;
     box-shadow: rgba(0, 0, 0, 0.14) 0 8px 10px 1px, rgba(0, 0, 0, 0.12) 0 3px 14px 2px, rgba(0, 0, 0, 0.2) 0 5px 5px -3px;
     width: 160px;
     right: 60px;
-    @extend .light-background;
 
     div {
       @apply h-12 pl-sm py-md text-body-l text-text-primary;
 
       &:hover {
-        @apply bg-grey-mid w-full;
         @extend .light-background;
+        @apply bg-grey-mid w-full;
       }
     }
   }
 
   .digit-search-add {
-    padding: 12px 16px;
     @extend .light-primary;
+    padding: 12px 16px;
     color: rgb(244, 119, 56);
     display: flex;
     cursor: pointer;
