+++ frontend/micro-ui/web/micro-ui-internals/packages/css/src/digitv2/components/breadCrumbV2.scss
@@ -1,4 +1,5 @@
 @import url("../index.scss");
+@import url("../typography.scss");
 
 .digit-bread-crumb {
   @extend .light-text-color-secondary;
@@ -22,6 +23,7 @@
   }
 
   span {
+    @extend .typography.text-body-s;
     cursor: pointer;
   }
 }
