+++ frontend/micro-ui/web/micro-ui-internals/packages/css/src/digitv2/components/cardV2.scss
@@ -1,8 +1,9 @@
 @import url("../index.scss");
+@import url("../typography.scss");
 
 .digit-card {
-  @apply bg-white m-sm px-md pt-md pb-lg shadow-card;
   @extend .light-paper-primary;
+  @apply bg-white m-sm px-md pt-md pb-lg shadow-card;
   border-radius: 4px;
   max-width: 960px;
 
@@ -23,7 +24,8 @@
 
   .digit-card-text {
     @extend .light-text-color-secondary;
-    @apply text-body-l align-middle text-left mb-lg;
+    @extend .typography.text-body-l;
+    @apply align-middle text-left mb-lg;
 
     span {
       @extend .light-text-color-primary;
@@ -32,7 +34,7 @@
 
   .digit-card-text-primary {
     @extend .light-primary;
-    @apply text-body-l;
+    @extend .typography.text-body-l;
   }
 
   .digit-card-text-button {
@@ -47,6 +49,7 @@
 
   .digit-card-label-error {
     @extend .alert-error;
+    @extend .typography.text-body-s;
     @apply block text-body-s text-error mb-md;
   }
 
@@ -79,6 +82,7 @@
     display: flex;
     align-items: center;
     .label {
+      @extend .typography.text-label;
       width: 30%;
       font-weight: 400;
       font-size: 16px;
@@ -118,7 +122,8 @@
 
   .digit-card-text {
     @extend .light-text-color-secondary;
-    @apply text-body-l align-middle text-left mb-lg;
+    @extend .typography.text-body-l;
+    @apply align-middle text-left mb-lg;
 
     span {
       @extend .light-text-color-primary;
@@ -181,11 +186,11 @@
 
 .card-emp {
   @extend .card;
+  @extend .light-background;
   padding-right: 0;
   padding-top: 0;
   padding-left: 0;
   @apply bg-white m-sm;
-  @extend .light-background;
 }
 
 .submit-bar {
@@ -248,7 +253,7 @@
   }
 
   .digit-employee-card {
-     @apply mb-md mx-md !important;
+    @apply mb-md mx-md !important;
     &.digit-filter {
       margin-left: auto;
       margin-right: auto;
@@ -344,6 +349,7 @@
 }
 
 .digit-card-label-error {
+  @extend .typography.text-body-s;
   @extend .alert-error;
   @apply block text-body-s text-error mb-md;
 }
