+++ frontend/micro-ui/web/micro-ui-internals/packages/css/src/digitv2/components/checkpointV2.scss
@@ -19,8 +19,8 @@
     @apply text-heading-s ml-md;
 
     p {
-      margin-top: 4px;
       @extend .light-text-color-secondary;
+      margin-top: 4px;
       @apply text-heading-s;
     }
   }
@@ -42,8 +42,8 @@
     @apply text-heading-s ml-md;
 
     p {
-      margin-top: 4px;
       @extend .light-text-color-secondary;
+      margin-top: 4px;
       @apply text-heading-s;
     }
   }
