+++ frontend/micro-ui/web/micro-ui-internals/packages/css/src/digitv2/components/dateWrapV2.scss
@@ -1,11 +1,14 @@
+@import url("../typography.scss");
+
 .digit-date-picker {
   @apply relative w-full cursor-pointer;
   .citizen {
     max-width: 540px;
   }
   .digit-employee-card-input {
-    @apply mb-lg pl-sm outline-none block border w-full h-10 border-input-border border-solid bg-white leading-10 text-form-field text-text-primary;
+    @extend .typography.text-body-s;
     @extend .light-input-border;
+    @apply mb-lg pl-sm outline-none block border w-full h-10 border-input-border border-solid bg-white leading-10 text-form-field text-text-primary;
     width: calc(100%-62px);
     padding-right: 9px;
   }
