+++ frontend/micro-ui/web/micro-ui-internals/packages/css/src/digitv2/components/loaderV2.scss
@@ -1,6 +1,7 @@
 @import url("../index.scss");
 
 .digit-screen-loader {
+  @extend .light-background;
   position: fixed;
   top: 0;
   left: 0;
@@ -10,7 +11,6 @@
   justify-content: center;
   align-items: center;
   z-index: 9999;
-  @extend .light-background;
   .digit-ldio-pjg92h09b2o div {
     position: absolute;
     width: 52px;
