+++ frontend/micro-ui/web/micro-ui-internals/packages/css/src/digitv2/components/paragraphV2.scss
@@ -1,0 +1,5 @@
+@import url("../typography.scss");
+
+.digit-paragraph {
+  @extend .typography.text-body-l;
+}
