+++ frontend/micro-ui/web/micro-ui-internals/packages/css/src/digitv2/components/textInputV2.scss
@@ -1,13 +1,14 @@
 @import url("../index.scss");
+@import url("../typography.scss");
 
 .digit-card-input-error {
   @apply mb-lg pl-sm block w-full h-10 outline-none border-2 border-error border-solid bg-white leading-10 text-form-field text-text-primary;
 }
 .digit-text-input {
   @apply relative w-full;
-  &.error {
+  &.digit-field-error {
     input {
-      border-color: red;
+      border: 1px solid theme(digitv2.alert.error);
     }
   }
   &.disabled {
@@ -28,13 +29,16 @@
 }
 
 .digit-employee-card-input {
+  @extend .typography.text-body-s;
   @apply mb-lg pl-sm outline-none block border w-full h-10 border-input-border border-solid bg-white leading-10 text-form-field text-text-primary;
 }
 .digit-employee-card-input:disabled {
+  @extend .typography.text-body-s;
   @apply border-grey-dark text-grey-dark !important;
   pointer-events: none !important;
 }
 .digit-employee-card-input--front {
+  @extend .typography.text-body-s;
   width: fit-content !important;
   display: flex;
   align-items: center;
@@ -43,6 +47,7 @@
   padding-right: 5px;
 }
 .digit-employee-card-input-error {
+  @extend .typography.text-body-s;
   @apply mb-lg pl-sm block w-full h-10 outline-none border-2 border-error border-solid bg-white leading-10 text-form-field text-text-primary;
 }
 
@@ -95,10 +100,12 @@
 }
 
 .digit-citizen-card-input {
+  @extend .typography.text-body-s;
   @apply mb-lg pl-sm outline-none block border w-full h-10 border-input-border border-solid bg-white leading-10 text-form-field text-text-primary;
 }
 
 .digit-citizen-card-input--front {
+  @extend .typography.text-body-s;
   width: fit-content !important;
   display: flex;
   align-items: center;
@@ -108,14 +115,15 @@
 }
 
 .digit-employee-card-input-error {
-  @apply mb-lg pl-sm block w-full h-10 outline-none border-2 border-error border-solid bg-white leading-10 text-form-field text-text-primary;
+  @extend .typography.text-body-s;
   @extend .light-input-border;
   @extend .alert-error;
+  @apply mb-lg pl-sm block w-full h-10 outline-none border-2 border-error border-solid bg-white leading-10 text-form-field text-text-primary;
 }
 .digit-card-input-error {
-  @apply mb-lg pl-sm block w-full h-10 outline-none border-2 border-error border-solid bg-white leading-10 text-form-field text-text-primary;
   @extend .light-input-border;
   @extend .alert-error;
+  @apply mb-lg pl-sm block w-full h-10 outline-none border-2 border-error border-solid bg-white leading-10 text-form-field text-text-primary;
 }
 .disabled {
   @apply border-grey-dark text-grey-dark !important;
