+++ frontend/micro-ui/web/micro-ui-internals/packages/css/src/digitv2/components/textareaV2.scss
@@ -1,9 +1,12 @@
+@import url("../typography.scss");
+
 .digit-card-textarea,
 .digit-employee-card-textarea {
-  @apply block outline-none mt-md mb-lg border w-full border-input-border border-solid bg-white h-24 text-form-field text-text-primary p-sm;
   @extend .light-input-border;
-  &.error {
-    border: 1px solid red;
+  @extend .typography.text-body-s;
+  @apply block outline-none mt-md mb-lg border w-full border-input-border border-solid bg-white h-24 text-form-field text-text-primary p-sm;
+  &.digit-field-error {
+    border: 1px solid theme(digitv2.alert.error);
   }
   &.disabled {
     @apply border-grey-dark text-grey-dark !important;
@@ -11,6 +14,7 @@
   }
 }
 .digit-employee-card-textarea {
+  @extend .typography.text-body-s;
   @apply w-full;
 }
 .digit-cell-text {
