+++ frontend/micro-ui/web/micro-ui-internals/packages/css/src/digitv2/index.scss
@@ -34,6 +34,7 @@
 @import url("./components/uploadFileV2.scss");
 @import url("./components/cardLabelV2.scss");
 @import url("./components/labelFieldPairV2.scss");
+@import url("./components/paragraphV2.scss");
 
 /* pages */
 @import url("./pages/employee/index.scss");
