+++ frontend/micro-ui/web/micro-ui-internals/packages/css/src/digitv2/typography.scss
@@ -39,6 +39,7 @@
       font-size: theme(digitv2.fontSize.heading-l.mobile);
     }
   }
+  
   &.text-heading-m {
     font-family: theme(digitv2.fontFamily.sans);
     font-style: theme(digitv2.fontStyle.normal);
@@ -57,6 +58,7 @@
       font-size: theme(digitv2.fontSize.heading-m.desktop);
     }
   }
+  
   &.text-heading-s {
     font-family: theme(digitv2.fontFamily.sans);
     font-style: theme(digitv2.fontStyle.normal);
@@ -75,6 +77,7 @@
       font-size: theme(digitv2.fontSize.heading-s.mobile);
     }
   }
+  
   &.text-heading-xs {
     @media (max-width: theme(digitv2.screens.mobile)) {
       font-size: theme(digitv2.fontSize.heading-xs.mobile);
@@ -85,6 +88,7 @@
       line-height: theme(digitv2.lineHeight.normal);
     }
   }
+  
   &.text-caption-l {
     font-family: theme(digitv2.fontFamily.sans);
     font-style: theme(digitv2.fontStyle.italic);
@@ -141,6 +145,7 @@
       font-size: theme(digitv2.fontSize.caption-s.desktop);
     }
   }
+  
   &.text-body-l {
     font-family: theme(digitv2.fontFamily.sans);
     font-style: theme(digitv2.fontStyle.italic);
@@ -161,6 +166,7 @@
       line-height: theme(digitv2.lineHeight.line-height-body-l.mobile);
     }
   }
+  
   &.text-body-s {
     font-family: theme(digitv2.fontFamily.sans);
     font-style: theme(digitv2.fontStyle.italic);
@@ -181,6 +187,7 @@
       line-height: theme(digitv2.lineHeight.line-height-body-s.mobile);
     }
   }
+  
   &.text-body-xs {
     font-family: theme(digitv2.fontFamily.sans);
     font-style: theme(digitv2.fontStyle.italic);
@@ -201,6 +208,7 @@
       line-height: theme(digitv2.lineHeight.line-height-body-xs.mobile);
     }
   }
+  
   &.text-label {
     font-family: theme(digitv2.fontFamily.sans);
     font-style: theme(digitv2.fontStyle.italic);
@@ -219,6 +227,7 @@
       font-size: theme(digitv2.fontSize.label.mobile);
     }
   }
+  
   &.text-link {
     font-family: theme(digitv2.fontFamily.sans);
     font-style: theme(digitv2.fontStyle.normal);
